refactor: implement modules-subtask
